Revised commands to include `net.probe on`
`net.probe on` is required to actively find other hosts in the local subnet to successfully arp spoof them as they join or leave.
Otherwise some traffic may be missed or not intercepted. At least when run as a service like this, `net.recon` and `arp.spoof` alone remain passive on the network.