Add arm64 and 20.04 support, additional functions
This update adds support for the latest Ubuntu 18.04 and 20.04 on both arm64 and x86_64.

It will need further testing on Raspberry Pi devices.
Some functions are still very basic and will need updated. (ManageCronTasks and Uninstall)

All components can be installed and managed individually for the required use case.
The functions are organized to (hopefully) allow easily updating and adding to what's available.